Add blacken script and format fixes
blacken.sh can be used by developers to run
format on the python source files in the repo.

The scope of files it runs on is limited at the moment
since running black on the jax files results in large
diffs which makes it hard to track changes from upstream.
Eventually when the PJRT stuff is fully integrated to the repo
we can run it on those. The jaxlib kernels is still an
open question.

The script will pass any args it receives to the inner black
command so args like --diff and --check work.