Explicitly pass the mi-250 label for nightly runs
Actions seems to be ignoring the default values for reusable jobs when
we call the job with the include option. This is fine for the ROCm
Jenkins build parameters, but causes the workflow to fail for the runner
label parameter. Just set the runner-label parameter explicitly for
builds with ROCm 6.4.1.